requirements.txt
google-api-python-client==2.186.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
requests==2.32.5
fuzzywuzzy==0.18.0
python-levenshtein==0.27.0
